Initial refactor based on family-tree-design-simplify. Still buggy.
